Add ugly hack for dashboard auto refresh
The MarkDownRenderer is a server component as it needs direct access
to the database for render context. A nice auto refresh would need to
be implemented client side as only part of the page needs to be re-rendered.

Switching MarkDownRenderer to be a client component is simply too much
work that I'm not willing to commit to so I've resorted back to this
ugly hack of server rendering a `<script>` tag that just refreshes
the entire page every minute.

Thank Sneh for all the work he did on this!